重构目录，新增甘特图粒度切换
