--- gantt_example.js
@@ -1,252 +1,0 @@
-let ganttInstance;
-const records = [
-  {
-    id: 1,
-    title: 'Task 1',
-    developer: 'liufangfang.jane@bytedance.com',
-    start: '2024-07-24',
-    end: '2024-07-26',
-    progress: 31,
-    priority: 'P0'
-  },
-  {
-    id: 2,
-    title: 'Task 2',
-    developer: 'liufangfang.jane@bytedance.com',
-    start: '07/24/2024',
-    end: '08/04/2024',
-    progress: 60,
-    priority: 'P0'
-  },
-  {
-    id: 3,
-    title: 'Task 3',
-    developer: 'liufangfang.jane@bytedance.com',
-    start: '2024-08-04',
-    end: '2024-08-04',
-    progress: 100,
-    priority: 'P1'
-  },
-  {
-    id: 4,
-    title: 'Task 4',
-    developer: 'liufangfang.jane@bytedance.com',
-    start: '2024-07-26',
-    end: '2024-07-28',
-    progress: 31,
-    priority: 'P0'
-  },
-  {
-    id: 5,
-    title: 'Task 5',
-    developer: 'liufangfang.jane@bytedance.com',
-    start: '2024-07-26',
-    end: '2024-07-28',
-    progress: 60,
-    priority: 'P0'
-  },
-  {
-    id: 6,
-    title: 'Task 6',
-    developer: 'liufangfang.jane@bytedance.com',
-    start: '2024-07-29',
-    end: '2024-08-11',
-    progress: 100,
-    priority: 'P1'
-  }
-];
-
-const columns = [
-  {
-    field: 'title',
-    title: 'title',
-    width: 'auto',
-    sort: true,
-    tree: true,
-    editor: 'input'
-  },
-  {
-    field: 'start',
-    title: 'start',
-    width: 'auto',
-    sort: true,
-    editor: 'date-input'
-  },
-  {
-    field: 'end',
-    title: 'end',
-    width: 'auto',
-    sort: true,
-    editor: 'date-input'
-  }
-];
-const option = {
-  overscrollBehavior: 'none',
-  records,
-  taskListTable: {
-    columns,
-    tableWidth: 250,
-    minTableWidth: 100,
-    maxTableWidth: 600,
-    theme: {
-      headerStyle: {
-        borderColor: '#e1e4e8',
-        borderLineWidth: 1,
-        fontSize: 18,
-        fontWeight: 'bold',
-        color: 'red',
-        bgColor: '#EEF1F5'
-      },
-      bodyStyle: {
-        borderColor: '#e1e4e8',
-        borderLineWidth: [1, 0, 1, 0],
-        fontSize: 16,
-        color: '#4D4D4D',
-        bgColor: '#FFF'
-      }
-    }
-    //rightFrozenColCount: 1
-  },
-  frame: {
-    outerFrameStyle: {
-      borderLineWidth: 2,
-      borderColor: '#e1e4e8',
-      cornerRadius: 8
-    },
-    verticalSplitLine: {
-      lineColor: '#e1e4e8',
-      lineWidth: 3
-    },
-    horizontalSplitLine: {
-      lineColor: '#e1e4e8',
-      lineWidth: 3
-    },
-    verticalSplitLineMoveable: true,
-    verticalSplitLineHighlight: {
-      lineColor: 'green',
-      lineWidth: 3
-    }
-  },
-  grid: {
-    verticalLine: {
-      lineWidth: 1,
-      lineColor: '#e1e4e8'
-    },
-    horizontalLine: {
-      lineWidth: 1,
-      lineColor: '#e1e4e8'
-    }
-  },
-  headerRowHeight: 40,
-  rowHeight: 40,
-  taskBar: {
-    startDateField: 'start',
-    endDateField: 'end',
-    progressField: 'progress',
-    resizable: true,
-    moveable: true,
-    hoverBarStyle: {
-      barOverlayColor: 'rgba(99, 144, 0, 0.4)'
-    },
-    labelText: '{title}  complete {progress}%',
-    labelTextStyle: {
-      fontFamily: 'Arial',
-      fontSize: 16,
-      textAlign: 'left',
-      textOverflow: 'ellipsis'
-    },
-    barStyle: {
-      width: 20,
-      /** 任务条的颜色 */
-      barColor: '#ee8800',
-      /** 已完成部分任务条的颜色 */
-      completedBarColor: '#91e8e0',
-      /** 任务条的圆角 */
-      cornerRadius: 8,
-      /** 任务条的边框 */
-      borderLineWidth: 1,
-      /** 边框颜色 */
-      borderColor: 'black'
-    }
-  },
-  timelineHeader: {
-    colWidth: 100,
-    backgroundColor: '#EEF1F5',
-    horizontalLine: {
-      lineWidth: 1,
-      lineColor: '#e1e4e8'
-    },
-    verticalLine: {
-      lineWidth: 1,
-      lineColor: '#e1e4e8'
-    },
-    scales: [
-      {
-        unit: 'week',
-        step: 1,
-        startOfWeek: 'sunday',
-        format(date) {
-          return `Week ${date.dateIndex}`;
-        },
-        style: {
-          fontSize: 20,
-          fontWeight: 'bold',
-          color: 'white',
-          strokeColor: 'black',
-          textAlign: 'right',
-          textBaseline: 'bottom',
-          backgroundColor: '#EEF1F5',
-          textStick: true
-          // padding: [0, 30, 0, 20]
-        }
-      },
-      {
-        unit: 'day',
-        step: 1,
-        format(date) {
-          return date.dateIndex.toString();
-        },
-        style: {
-          fontSize: 20,
-          fontWeight: 'bold',
-          color: 'white',
-          strokeColor: 'black',
-          textAlign: 'right',
-          textBaseline: 'bottom',
-          backgroundColor: '#EEF1F5'
-        }
-      }
-    ]
-  },
-  markLine: [
-    {
-      date: '2024/8/02',
-      scrollToMarkLine: true,
-      position: 'left',
-      style: {
-        lineColor: 'red',
-        lineWidth: 1
-      }
-    }
-  ],
-  rowSeriesNumber: {
-    title: '行号',
-    dragOrder: true,
-    headerStyle: {
-      bgColor: '#EEF1F5',
-      borderColor: '#e1e4e8'
-    },
-    style: {
-      borderColor: '#e1e4e8'
-    }
-  },
-  scrollStyle: {
-    scrollRailColor: 'RGBA(246,246,246,0.5)',
-    visible: 'scrolling',
-    width: 6,
-    scrollSliderCornerRadius: 2,
-    scrollSliderColor: '#5cb85c'
-  }
-};
-ganttInstance = new VTableGantt.Gantt(document.getElementById(CONTAINER_ID), option);
-window['ganttInstance'] = ganttInstance;
