+++ main/store.js
@@ -1,0 +1,4 @@
+const Store = require('electron-store');
+const store = new Store();
+
+module.exports = store;
