+++ renderer/index.html
@@ -6,13 +6,14 @@
   <title>EasyMemo</title>
   <link rel="stylesheet" href="styles.css">
   <link href="https://fonts.googleapis.com/css2?family=Noto+Sans+SC:wght@400;500;700&display=swap" rel="stylesheet">
-  <script src="node_modules/@visactor/vtable-gantt/dist/vtable-gantt.min.js"></script>
+  <script src="../node_modules/@visactor/vtable-gantt/dist/vtable-gantt.min.js"></script>
+  <script src="gantt.js"></script>
 </head>
 <body>
   <div class="container">
     <header class="app-header">
       <div class="logo">
-        <img src="icon.svg" alt="memo" width="32" height="32">
+        <img src="../image/icon.svg" alt="memo" width="32" height="32">
         <h1>EasyMemo</h1>
       </div>
       <div class="app-menu">
@@ -40,6 +41,7 @@
               <th>序号</th>
               <th>项目名称</th>
               <th>开始时间</th>
+              <th>结束时间</th>
               <th>最新修改时间</th>
               <th>完成事项数</th>
               <th>未完成事项数</th>
@@ -84,7 +86,18 @@
         </div>
       </div>
       <div id="gantt-view" class="view-content">
-        <div id="gantt-container" style="width: 100%; height: 600px;"></div>
+        <div class="gantt-toolbar">
+          <div class="gantt-toolbar-left">
+            <label for="time-scale-select">时间粒度：</label>
+            <select id="time-scale-select" class="time-scale-selector">
+              <option value="week">周</option>
+              <option value="month" selected>月</option>
+              <option value="quarter">季</option>
+              <option value="year">年</option>
+            </select>
+          </div>
+        </div>
+        <div id="gantt-container" style="width: 100%; height: calc(100% - 50px);"></div>
       </div>
     </div>
   </div>
@@ -101,6 +114,14 @@
         <div class="form-group">
           <label for="project-name">项目名称</label>
           <input type="text" id="project-name" required>
+        </div>
+        <div class="form-group">
+          <label for="project-start-date">开始时间</label>
+          <input type="date" id="project-start-date">
+        </div>
+        <div class="form-group">
+          <label for="project-end-date">结束时间</label>
+          <input type="date" id="project-end-date">
         </div>
         <div class="form-actions">
           <button type="submit">保存</button>
